Installer updates

+++ vim/install.sh
@@ -1,1 +1,5 @@
-ln -s ~/.dotfiles/vim/.vimrc ~/.vimrc
+#!/bin/bash
+
+if [ ! -f ~/.vimrc ]; then
+	ln -s ~/.dotfiles/vim/.vimrc ~/.vimrc
+fi
